remove some duplicated portions
