+++ callback_common.go
@@ -3,7 +3,10 @@
 
 package purego
 
-import "unsafe"
+import (
+	"runtime"
+	"unsafe"
+)
 
 // maxCb is the maximum number of callbacks
 // only increase this if you have added more to the callbackasm function
@@ -13,3 +16,27 @@ const maxCB = 2000
 var callbackasmABI0 uintptr
 
 const ptrSize = unsafe.Sizeof((*int)(nil))
+
+// callbackasmAddr returns address of runtime.callbackasm
+// function adjusted by i.
+// On x86 and amd64, runtime.callbackasm is a series of CALL instructions,
+// and we want callback to arrive at
+// correspondent call instruction instead of start of
+// runtime.callbackasm.
+// On ARM, runtime.callbackasm is a series of mov and branch instructions.
+// R12 is loaded with the callback index. Each entry is two instructions,
+// hence 8 bytes.
+func callbackasmAddr(i int) uintptr {
+	var entrySize int
+	switch runtime.GOARCH {
+	default:
+		panic("unsupported architecture")
+	case "386", "amd64":
+		entrySize = 5
+	case "arm", "arm64":
+		// On ARM, and ARM64 each entry is a MOV instruction
+		// followed by a branch instruction
+		entrySize = 8
+	}
+	return callbackasmABI0 + uintptr(i*entrySize)
+}
